pyproject.toml
[project]
name = "tg-bot-pattern"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.20.0.post0",
    "apscheduler>=3.11.0",
    "cache3>=0.4.3",
    "loguru>=0.7.3",
    "tortoise-orm>=0.25.0",
]
